Major rewrite
The issue before was that the parent path used for iterative dumping was
incorrect, causing criu to perform "full" predumps (i.e. it always
dumped the entire memory state). This was noticed when running the
benchmarking scripts and noticing that the size of the predump was often
at least as large as the size of the full dumps.

Now, dumps are organized into linked lists called chains. The system
hold two of these in memory, called the previous chain and current
chain.
The previous chain is always completed (i.e. no more dumps will
be added to it) and represents the latest state from which a process a
can be restored.
The current chain is not completed and is the chain to which the system
adds new dumps, either via regular dumps or when migrating.
When a migration is requested a last predump and full dump is added to
the end of the current chain, and a migration RPC containing the name of
all the dumps belonging to the chain from which to restore is sent.
During failover, the source can not tell the other host which dumps
belong to the chain from which to restore, instead it has to figure it
out itself. It does this by finding the latest dump directory (as
before) and then recursively iterates over the dump directories using
the parent symlink from said directories.
The last node of a previous chain will always be a full dump.
The length of a chain is determined via the `CHAIN_LENGTH` environment
variable, and the environment variable `FULLDUMP_FREQ` has been removed.

Now, the user is told if no recoverable dump directory was found.

Replaces dump offset with chain length.